+++ gpiozero/pins/pigpio.py
@@ -103,7 +103,7 @@ class PiGPIOFactory(PiFactory):
         bug in our pin implementation). A workaround for now is simply to
         restart the :command:`pigpiod` daemon.
 
-    .. _pigpio: https://pypi.org/project/pigpio/
+    .. _pigpio: http://abyz.me.uk/rpi/pigpio/
     """
     def __init__(self, host=None, port=None):
         super(PiGPIOFactory, self).__init__()
@@ -185,7 +185,7 @@ class PiGPIOPin(PiPin):
     Extends :class:`~gpiozero.pins.pi.PiPin`. Pin implementation for the
     `pigpio`_ library. See :class:`PiGPIOFactory` for more information.
 
-    .. _pigpio: http://abyz.co.uk/rpi/pigpio/
+    .. _pigpio: http://abyz.me.uk/rpi/pigpio/
     """
     _CONNECTIONS = {} # maps (host, port) to (connection, pi_info)
     GPIO_FUNCTIONS = {
@@ -348,7 +348,7 @@ class PiGPIOHardwareSPI(SPI, Device):
     Hardware SPI implementation for the `pigpio`_ library. Uses the ``spi_*``
     functions from the pigpio API.
 
-    .. _pigpio: http://abyz.co.uk/rpi/pigpio/
+    .. _pigpio: http://abyz.me.uk/rpi/pigpio/
     """
     def __init__(self, factory, port, device):
         self._port = port
@@ -450,7 +450,7 @@ class PiGPIOSoftwareSPI(SPI, Device):
     Software SPI implementation for the `pigpio`_ library. Uses the ``bb_spi_*``
     functions from the pigpio API.
 
-    .. _pigpio: http://abyz.co.uk/rpi/pigpio/
+    .. _pigpio: http://abyz.me.uk/rpi/pigpio/
     """
     def __init__(self, factory, clock_pin, mosi_pin, miso_pin, select_pin):
         self._closed = True